IO_URING I/O Support for Linux >= 5.6 Added support for seastar as server backend Fixed issues in SELECT/POLL I/O Fixed issues in EVENT_PORTS/DEVPOLL for Solaris Fixed issues in SSL/TLS support Added generic I/O event notification callback based loop support Updated README Serialization performance improvement Version bumped to 5.3 [docker_push][ci skip]

+++ src/modules/cache/memory/MemoryCacheImpl.cpp
@@ -153,7 +153,7 @@ bool MemoryCacheImpl::flushAll() {
 
 void MemoryCacheImpl::clean() {
 	MemoryCacheConnectionPool* p = (MemoryCacheConnectionPool*)pool;
-	while(p->internalMap.size()>p->size) {
+	while((int)p->internalMap.size()>p->size) {
 		auto last_it = p->lrul.end(); last_it --;
 		p->internalMap.erase(last_it->first);
 		p->lrul.pop_back();
